[WEEK7] Thread 실행법 Runnable, new Thread, Lamda
