+++ week6/Thread/src/dev/syntax/step01createthread/Main.java
@@ -1,3 +1,5 @@
+package dev.syntax.step01createthread;
+
 public class Main {
 	public static void main(String[] args) {
 		System.out.println("Hello world!");
