+++ week6/Thread/src/step02runningthread/Ex01.java
@@ -1,0 +1,7 @@
+package step02runningthread;
+
+public class Ex01 {
+	public static void main(String[] args) {
+
+	}
+}
